+++ Tests/Editor/ShadersAnalysisTests.cs
@@ -6,6 +6,7 @@ using NUnit.Framework;
 using Unity.ProjectAuditor.Editor;
 using Unity.ProjectAuditor.Editor.Diagnostic;
 using Unity.ProjectAuditor.Editor.Modules;
+using Unity.ProjectAuditor.Editor.TestUtils;
 using Unity.ProjectAuditor.Editor.Utils;
 using UnityEditor;
 using UnityEditor.Build;
@@ -20,14 +21,14 @@ namespace Unity.ProjectAuditor.EditorTests
         const string k_ShaderName = "Custom/MyTestShader,1"; // comma in the name for testing purposes
 
 #pragma warning disable 0414
-        TempAsset m_ShaderResource;
-        TempAsset m_PlayerLogResource;
-        TempAsset m_PlayerLogWithNoCompilationResource;
-        TempAsset m_ShaderWithErrorResource;
-        TempAsset m_EditorShaderResource;
+        TestAsset m_ShaderResource;
+        TestAsset m_PlayerLogResource;
+        TestAsset m_PlayerLogWithNoCompilationResource;
+        TestAsset m_ShaderWithErrorResource;
+        TestAsset m_EditorShaderResource;
 
-        TempAsset m_ShaderUsingBuiltInKeywordResource;
-        TempAsset m_SurfShaderResource;
+        TestAsset m_ShaderUsingBuiltInKeywordResource;
+        TestAsset m_SurfShaderResource;
 #pragma warning restore 0414
 
         const string s_KeywordName = "DIRECTIONAL";
@@ -67,7 +68,7 @@ namespace Unity.ProjectAuditor.EditorTests
         [OneTimeSetUp]
         public void SetUp()
         {
-            m_ShaderResource = new TempAsset("Resources/MyTestShader.shader", @"
+            m_ShaderResource = new TestAsset("Resources/MyTestShader.shader", @"
             Shader ""Custom/MyTestShader,1""
             {
                 SubShader
@@ -150,7 +151,7 @@ namespace Unity.ProjectAuditor.EditorTests
                 }
             }");
 
-            m_PlayerLogResource = new TempAsset("player.log", @"
+            m_PlayerLogResource = new TestAsset("player.log", @"
 02-10 17:36:20.945  6554  6816 D Unity   : Compiled shader: Custom/MyTestShader,1, pass: MyTestShader/Pass, stage: vertex, keywords <no keywords>
 02-10 17:36:20.945  6554  6816 D Unity   : Compiled shader: Custom/MyTestShader,1, pass: MyTestShader/Pass, stage: fragment, keywords <no keywords>
 02-10 17:36:20.945  6554  6816 D Unity   : Compiled shader: Custom/MyTestShader,1, pass: MyTestShader/Pass, stage: vertex, keywords KEYWORD_A
@@ -161,20 +162,20 @@ namespace Unity.ProjectAuditor.EditorTests
             ");
 
 
-            m_PlayerLogWithNoCompilationResource = new TempAsset("player_with_no_compilation.log", string.Empty);
+            m_PlayerLogWithNoCompilationResource = new TestAsset("player_with_no_compilation.log", string.Empty);
 
 #if UNITY_2021_1_OR_NEWER
             UnityEngine.TestTools.LogAssert.ignoreFailingMessages = true; // workaround for 2021.x failure
 #endif
 
 #if UNITY_2019_1_OR_NEWER
-            m_ShaderWithErrorResource = new TempAsset("Resources/ShaderWithError.shader", @"
+            m_ShaderWithErrorResource = new TestAsset("Resources/ShaderWithError.shader", @"
             Sader ""Custom/ShaderWithError""
             {
             }");
 #endif
 
-            m_ShaderUsingBuiltInKeywordResource = new TempAsset("Resources/ShaderUsingBuiltInKeyword.shader", @"
+            m_ShaderUsingBuiltInKeywordResource = new TestAsset("Resources/ShaderUsingBuiltInKeyword.shader", @"
 Shader ""Custom/ShaderUsingBuiltInKeyword""
             {
                 Properties
@@ -228,7 +229,7 @@ Shader ""Custom/ShaderUsingBuiltInKeyword""
             }
             ");
 
-            m_SurfShaderResource = new TempAsset("Resources/MySurfShader.shader", @"
+            m_SurfShaderResource = new TestAsset("Resources/MySurfShader.shader", @"
 Shader ""Custom/MySurfShader""
             {
                 Properties
@@ -274,7 +275,7 @@ Shader ""Custom/MySurfShader""
             }
 ");
 
-            m_EditorShaderResource = new TempAsset("Editor/MyEditorShader.shader", @"
+            m_EditorShaderResource = new TestAsset("Editor/MyEditorShader.shader", @"
 Shader ""Custom/MyEditorShader""
             {
                 Properties